Rename task to thrift.generate
